requirements.txt
Django==1.7.1
Jinja2==2.7.3
MarkupSafe==0.23
Pillow==2.6.1
argparse==1.2.1
blessings==1.6
coverage==3.7.1
html2text==2014.12.24
ipdb==0.8
ipython==2.3.1
lxml==3.4.1
nose==1.3.4
nose-progressive==1.5.1
pyenchant==1.6.6
python-termstyle==0.1.10
redis==2.10.3
rednose==0.4.1
selenium==2.44.0
splinter==0.7.0
wsgiref==0.1.2
xamcheck-utils==0.0.8
